Organized theme keys, removed unused keys (#744)
* Organized theme keys, removed unused keys

After creating a few themes, I've noticed the previous base00...base0L naming to be a bit confusing. In part this is because the order of keys is so arbitrary after 07. This commit groups keys into 3 main categories so that, hopefully, it's easier for folks to create new themes.

I also noticed that 2 of the previous 00...07 keys weren't being used. I've deleted them.

* Merged master and updated Apathy theme

* Lintix